add to wishlist and location changes

+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/foods/foodsShopPreview/modelShopHome/Option.kt
@@ -1,9 +1,10 @@
 package com.teamx.hatlyUser.ui.fragments.foods.foodsShopPreview.modelShopHome
 import androidx.annotation.Keep
+
 @Keep
 data class Option(
     val _id: String,
     val name: String,
-    val price: Int,
-    val salePrice: Int
+    val price: Double,
+    val salePrice: Double
 )
